moved the nti responses to contracts
