+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Controllers/NTIWarningLetterControllerTests.cs
@@ -1,7 +1,6 @@
 ﻿using ConcernsCaseWork.API.Contracts.Common;
 using ConcernsCaseWork.API.Contracts.NtiWarningLetter;
 using ConcernsCaseWork.API.Features.NTIWarningLetter;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.WarningLetter;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Models;
 using FizzWare.NBuilder;
