+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Controllers/NoticeToImproveControllerTests.cs
@@ -1,7 +1,6 @@
 ﻿using ConcernsCaseWork.API.Contracts.Common;
 using ConcernsCaseWork.API.Contracts.NoticeToImprove;
 using ConcernsCaseWork.API.Features.NoticeToImprove;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.NoticeToImprove;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Models;
 using FizzWare.NBuilder;
