+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Factories/NTIUnderConsiderationFactoryTests.cs
@@ -1,6 +1,5 @@
 ﻿using ConcernsCaseWork.API.Contracts.NtiUnderConsideration;
 using ConcernsCaseWork.API.Factories.CaseActionFactories;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.UnderConsideration;
 using ConcernsCaseWork.Data.Models;
 using FizzWare.NBuilder;
 using FluentAssertions;
