+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Integration/NtiUnderConsiderationIntegrationTests.cs
@@ -1,9 +1,8 @@
 ﻿using AutoFixture;
 using ConcernsCaseWork.API.Contracts.Case;
 using ConcernsCaseWork.API.Contracts.Common;
+using ConcernsCaseWork.API.Contracts.NoticeToImprove;
 using ConcernsCaseWork.API.Contracts.NtiUnderConsideration;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.NoticeToImprove;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.UnderConsideration;
 using ConcernsCaseWork.API.Tests.Fixtures;
 using ConcernsCaseWork.API.Tests.Helpers;
 using FizzWare.NBuilder;
