+++ ConcernsCaseWork/ConcernsCaseWork.API/Factories/CaseActionFactories/NTIWarningLetterFactory.cs
@@ -1,5 +1,4 @@
 ﻿using ConcernsCaseWork.API.Contracts.NtiWarningLetter;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.WarningLetter;
 using ConcernsCaseWork.Data.Models;
 
 namespace ConcernsCaseWork.API.Factories.CaseActionFactories
