+++ ConcernsCaseWork/ConcernsCaseWork.API/Factories/CaseActionFactories/NoticeToImproveFactory.cs
@@ -1,5 +1,4 @@
 ﻿using ConcernsCaseWork.API.Contracts.NoticeToImprove;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.NoticeToImprove;
 using ConcernsCaseWork.Data.Models;
 
 namespace ConcernsCaseWork.API.Factories.CaseActionFactories
