+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/ConcernsStatus/ConcernsStatusController.cs
@@ -1,6 +1,6 @@
 using ConcernsCaseWork.API.Contracts.Common;
 using ConcernsCaseWork.API.Contracts.Concerns;
-using ConcernsCaseWork.API.ResponseModels;
+using ConcernsCaseWork.API.Factories;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ConcernsCaseWork.API.Features.ConcernsStatus
