+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/MeansOfReferral/IndexConcernsMeansOfReferrals.cs
@@ -1,6 +1,5 @@
 ﻿using ConcernsCaseWork.API.Contracts.Concerns;
 using ConcernsCaseWork.API.Factories;
-using ConcernsCaseWork.API.ResponseModels;
 using ConcernsCaseWork.Data.Gateways;
 
 namespace ConcernsCaseWork.API.Features.MeansOfReferral
