+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/NTIUnderConsideration/GetNTIUnderConsiderationById.cs
@@ -1,9 +1,7 @@
-﻿using ConcernsCaseWork.API.Factories;
+﻿using ConcernsCaseWork.API.Contracts.NtiUnderConsideration;
 using ConcernsCaseWork.API.Factories.CaseActionFactories;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.UnderConsideration;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Gateways;
-using ConcernsCaseWork.Data.Models;
 
 namespace ConcernsCaseWork.API.Features.NTIUnderConsideration
 {
