+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/NTIWarningLetter/GetNTIWarningLetterById.cs
@@ -1,11 +1,7 @@
-﻿using Azure.Core;
-using ConcernsCaseWork.API.Contracts.ResponseModels.Concerns.Decisions;
-using ConcernsCaseWork.API.Factories;
+﻿using ConcernsCaseWork.API.Contracts.NtiWarningLetter;
 using ConcernsCaseWork.API.Factories.CaseActionFactories;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.WarningLetter;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Gateways;
-using System;
 
 namespace ConcernsCaseWork.API.Features.NTIWarningLetter
 {
