+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/NTIWarningLetter/NTIWarningLetterController.cs
@@ -1,6 +1,5 @@
 ﻿using ConcernsCaseWork.API.Contracts.Common;
 using ConcernsCaseWork.API.Contracts.NtiWarningLetter;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.WarningLetter;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Models;
 using Microsoft.AspNetCore.Mvc;
